# perf: return least complex marker when building intersection