This is analogous to what we do when building the union. Although this means extra work, it pays off if the resulting marker is used in another marker operation.

Relates to: python-poetry/poetry#9956

The downstream failures are caused by cosmetic changes. The tests are fixed in python-poetry/poetry#10077. We should probably merge this PR with failing downstream tests and switch from the released poetry-core version to the main branch afterwards.

<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes. But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [ ] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

<!--
**Note**: All Pull Requests must be based on the `main` branch.

[If you have *any* questions to *any* of the points above, just **submit and ask**!  This checklist is here to *help* you, not to deter you from contributing!
-->

After the simplified repro reported in https://github.com/python-poetry/poetry/issues/9956#issuecomment-2573922549 is almost as fast as before (see #818), I took a look at the example from https://github.com/python-poetry/poetry/issues/9956#issuecomment-2582451109, which is still significantly slower. With this PR, it is about as fast as before:

poetry version|lock (--regenerate)|lock (--no-update)
---|---|---
1.8|178 s|15 s
main|>600 s|120 s
PR|151 s|19 s

## Summary by Sourcery

Optimize marker intersection performance by returning the least complex marker representation. This addresses performance regressions observed when intersecting markers.

Bug Fixes:
- Fixed a performance issue where marker intersection could be significantly slower in certain cases.

Tests:
- Added tests to verify the performance improvements of marker intersection.

## Summary by Sourcery

Optimize marker intersection performance by returning the least complex marker representation.

Bug Fixes:
- Fixed a performance issue where marker intersection could be significantly slower in certain cases.

Tests:
- Added tests to verify the performance improvements.